% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{authenticate}
\alias{authenticate}
\title{authenticate dwhr session.}
\usage{
authenticate(session)
}
\value{
als geauthenticiteerd dan TRUE ander FALSE.
}
\description{
authenticeren van de gebruiker. Wordt gestuurd door de globale variable securityModel.
}
