% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwhr.R
\name{changeDtFgStyle}
\alias{changeDtFgStyle}
\alias{changeDtBgStyle}
\alias{setDtVisible}
\title{Wijzig stijl of zichtbaarheid van een kolom in een dataTable presentatie.}
\usage{
changeDtFgStyle(env, dim, pres, viewColumn, fgStyle = NULL)

changeDtBgStyle(env, dim, pres, viewColumn, bgStyle = NULL)

setDtVisible(env, dim, pres, viewColumn, visible)
}
\arguments{
\item{env}{sterschema object, gecreeerd met \code{\link{new.star}}.}

\item{dim}{string, dimView id gecreeerd met \code{\link{addDimView}}.}

\item{pres}{string, naam van de dataTable presentatie zoals gecreerd met \code{\link{addPresentation}} en
waarin te wijzigen kolom voorkomt.}

\item{viewColumn}{string, identificatie van de kolom. Dit is de naam zoals in eerste instantie meegegeven aan \code{\link{addMeasure}} 
of aan \code{\link{addMeasureDerrived}} of zoals die in deze functies wordt afgeleid en die bovendien voorkomt in eerder opgegeven presentatie.}

\item{fgStyle}{list, specificatie van de nieuwe foreground style van deze viewColumn. List heeft zelfde structuur als bij de bgStyle parameter.}

\item{bgStyle}{list, specificatie van de nieuwe background style van deze viewColumn. List kent volgende waarden:
\itemize{
   \item cuts: numeric, vector van waarden waar de kolomwaarde over verdeeld wordt.
   \item levels: numeric, vector van waarden waar de kolomwaarde mee overeen moet stemmen.  Of cuts wordt opgegven of levels, 1 van beiden. 
   \item values: character, vector met kleuren waar de kolomwaarde op afgebeeld wordt via ofwel cuts ofwel levels. In het geval van cuts moet de vector van kleuren 
1 langer zijn dan de vector cuts. Bij gebruik van levels moet de vector van kleuren even lang zijn als de levels-vector.
   \item valueColumn: Normaal wordt de waarde uit de viewColumn zelf gehaald, de waarde kan echter ook uit een andere viewColumn komen.
}}

\item{visible}{boolean, controleert de zichtbaarheid van een datatable kolom}
}
\description{
Wijzig stijl of zichtbaarheid van een kolom in een dataTable presentatie.
}
