% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwhr.R
\name{changeFormatMeasure}
\alias{changeFormatMeasure}
\title{Wijzig format van bestaande meetwaarde}
\usage{
changeFormatMeasure(env, dim, viewColumn, format)
}
\arguments{
\item{env}{sterschema object, gecreeerd met \code{\link{new.star}}.}

\item{dim}{string, dimView id gecreeerd met \code{\link{addDimView}}.}

\item{viewColumn}{character, naam van de technische meetwaarde.}

\item{format}{character, nieuw te gebruiken format.
\itemize{
  \item standard: waarde getoond as is.
  \item integer: afronding naar dichtsbijzijnde integer waarde.
  \item euro: afronding naar geheel getal met euro teken ervoor.
  \item euro2: idem maar dan afgerond op 2 decimalen.
  \item keuro: waarde wordt gedeeld door duizend en afgrond naar geheel getal. Euro teken wordt voor het resultaat geplaatst.
  \item perc: afronding naar geheel getal + %
  \item perc1: afronding op 1 decimaal + %
  \item perc2: afronding op 2 decimalen + %
  \item decimal1: waarde afgerond op 1 decimaal
  \item decimal2: waarde afgerond op 2 decimalen
  \item decimal3: waarde afgerond op 3 decimalen
}}
}
\description{
Pas het format aan van een meetwaarde uit een dimView. Er wordt gerefereerd naar de meetwaarde via de 
gepresenteerde naam (as).
}
