% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwhr.R
\name{dimChangeState}
\alias{dimChangeState}
\title{Aanpassen state van dimView}
\usage{
dimChangeState(env, dim, newState)
}
\arguments{
\item{env}{sterschema-object, gemaakt met \code{\link{new.star}}}

\item{dim}{string, dimView id gecreeerd met \code{\link{addDimView}}.}

\item{newState}{string, nieuwe toestand voor dimView}
}
\description{
met deze functie kan de toestand van een dimView aangepast worden. Toegestane toestanden (state) zijn:
\itemize{
 \item enabled: dimView heeft filterend effect op de feiten en is zichtbaar.
 \item hidden: dimView niet zichtbaar, maar heeft nog steeds een filterend effect op de feiten.
 \item disabled: dimView heeft geen filterend effect en is niet zichtbaar.
}
}
