% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwhr-package.R
\docType{package}
\name{dwhr-package}
\alias{dwhr}
\alias{dwhr-package}
\title{dwhr: Interactive Shiny Dashboards Over Star-Schema Data Warehouse Data}
\description{
Provides a framework for building interactive Shiny dashboards over
star-schema data warehouse data. Users construct a `star` object from a
facts table, dimension views, measures, and presentation specifications;
the package handles the reactive plumbing between drillable hierarchical
dimensions and DataTable, Highcharts, and form control presentations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/howardrcc/dwhr}
  \item Report bugs at \url{https://github.com/howardrcc/dwhr/issues}
}

}
\author{
\strong{Maintainer}: Howard Ching Chung \email{howardchingchung@protonmail.com}

Authors:
\itemize{
  \item Pieter Timmerman (original author, in memoriam)
}

}
\keyword{internal}
