% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{getDbHandle}
\alias{getDbHandle}
\title{get database-handle voor omgeving}
\usage{
getDbHandle(omg, db = "R")
}
\arguments{
\item{omg}{string, naam van te benaderen omgeving.}

\item{db}{string, database naam, default R.}
}
\value{
database-handle
}
\description{
database-handle kan worden gebruikt voor sql-queries op de betreffende omgeving. Het bestand dbCred.rds moet aanwezig
zijn in de data directory van de shiny app.
}
