% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{getDimUI}
\alias{getDimUI}
\title{Create dimView object in the UI}
\usage{
getDimUI(dim, useWellPanel = FALSE, maxHeight = NULL,
  overflowX = "hidden")
}
\arguments{
\item{dim}{string, unique id of the dimView. This id should correspond to the parameter \code{dim} in
\code{\link{addDimView}}}

\item{useWellPanel}{boolean, if TRUE a wellPanel is drawn around the dimView object}

\item{maxHeight}{an integer, maxheight in pixels for this dimView}

\item{overflowX}{?}
}
\description{
Create html for dimView object in the shiny app's UI.
}
