% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwhr.R
\name{getFacts}
\alias{getFacts}
\alias{getCSV}
\title{Inlezen csv bestand in een data.table of data.frame.}
\usage{
getFacts(
  file,
  key = NULL,
  useRDS = TRUE,
  sep = ";",
  fileEncoding = "UTF-8-BOM",
  as.df = FALSE,
  header = FALSE,
  ...
)

getCSV(...)
}
\arguments{
\item{file}{string, pad van bron bestand.}

\item{key}{character, optionele vector met keys om toe te passen op het resultaat. Is voor performance redenen}

\item{useRDS}{boolean, als TRUE dan wordt rds aangemaakt en gelezen, anders altijd lezen vanuit csv. Het rds bestand wordt 
opgeslagen in de tmp directory van de shiny-app. De tmp directory moet dus wel bestaan.}

\item{sep}{string, te gebruiken separator}

\item{fileEncoding}{string, te gebruiken encoding, default UTF-8-BOM ivm Byte Order Mark in windows bestanden}

\item{as.df}{boolean, als TRUE maak dan een data.frame aan ipv een data.table}

\item{...}{overige parameters zoals geaccepteerd door read.csv}
}
\description{
Bestand kan na initieele load opgeslagen worden als rds.
Dit rds bestand wordt dan de volgende keer als bron gebruikt (veel sneller als inlezen csv)
}
\details{
Als csv bestand een recentere datum heeft dan bestaande rds, dan wordt het csv bestand weer als bron 
genomen. \code{getCSV} is een alias voor \code{getFacts}.
}
