% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwhr.R
\name{setColumnName}
\alias{setColumnName}
\title{Pas de gepresenteerde kolomnaam aan van een meetwaarde uit een dimView.}
\usage{
setColumnName(env, dim, colFrom, viewColFrom = NULL, colTo)
}
\arguments{
\item{env}{sterschema object, gecreeerd met \code{\link{new.star}}.}

\item{dim}{string, dimView id gecreeerd met \code{\link{addDimView}}.}
}
\value{
gewijzigd sterschema-object.
}
\description{
Er wordt gerefereerd naar de meetwaarde via de gepresenteerde naam (colFrom, colTo). Alternatieve selectie van 
de te wijzigen kolom kan ook via de viewColumn (viewColFrom)
}
