% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwhr.R
\name{setOrdering}
\alias{setOrdering}
\title{Wijzig kolom(men) waarop dimView in de UI gesorteerd wordt.}
\usage{
setOrdering(env, dim, as, sort, as2 = NULL)
}
\arguments{
\item{env}{sterschema object, gecreeerd met \code{\link{new.star}}.}

\item{dim}{string, dimView id gecreeerd met \code{\link{addDimView}}.}

\item{as}{string, naam van de eerste kolom waarop gesorteerd moet worden}

\item{sort}{string, richting van sorteren. Mogelijk waarden: 
\itemize{
  \item asc: oplopend sorteren
  \item desc: aflopend sorteren
  \item LH: oplopend sorteren
  \item HL: aflopend sorteren
}}

\item{as2}{string, naam van de tweede kolom waarop gesorteerd moet worden}
}
\description{
Pas de sorteer kolommen van dimView aan. Er wordt gerefereerd naar de kolom via de 
gepresenteerde naam (as).
}

