% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwhr.R
\name{setSelection}
\alias{setSelection}
\title{Zet programmatisch de selectie van een dimView.}
\usage{
setSelection(
  env,
  dim,
  sel,
  source = "setSelection",
  dimRefresh = TRUE,
  selectChange = TRUE
)
}
\arguments{
\item{env}{sterschema object, gecreeerd met \code{\link{new.star}}.}

\item{dim}{string, dimView id gecreeerd met \code{\link{addDimView}}.}

\item{sel}{data.frame, data.frame met te selecteren items. Format:
\itemize{
    \item level, level van te selecteren items.
    \item label, label van te selecteren items.
    \item parent, parent-label van de te selecteren items.
}}

\item{source}{string, extra info over de bron van de selectieverandering.}

\item{dimRefresh}{boolean, indicatie of UI-output bijgewerkt moet worden of niet.}
}
\description{
De selectie van een dimView kan enerzijds via de UI ingesteld worden, anderszijds is het vaak nodig 
de selectie op programmatische wijze aan te passen, dit kan via setSelection.
}
