% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{getDimUI}
\alias{getDimUI}
\title{Creeer dimView object in the UI}
\usage{
getDimUI(dim, useWellPanel = FALSE, maxHeight = NULL,
  overflowX = "hidden")
}
\arguments{
\item{dim}{string, uniek id van the dimView. Dit id moet overeenkomen met de parameter \code{dim} in
\code{\link{addDimView}}}

\item{useWellPanel}{boolean, Als TRUE: er wordt een wellPanel om het dimView object getekend}

\item{maxHeight}{an integer, maximum hoogte in pixels voor dit dimView object}

\item{overflowX}{string, css overflow propery in X richting, bepaalt of er wel of niet een horizontale scrollbar getoond wordt bij overflow.}
}
\description{
Creeer de html voor het dimView object in de shiny app's UI.
}
