% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{getDimUI}
\alias{getDimUI}
\title{Creeer dimView object in the UI}
\usage{
getDimUI(starId, dim, skipTopRow = FALSE, maxHeight = NULL,
  overflowX = "hidden", accordion = FALSE)
}
\arguments{
\item{starId, }{uniek id van het sterschema. Dit id moet overeenkomen met de parameter \code{starId} in \code{\link{new.star}}}

\item{dim}{string, uniek id van the dimView. Dit id moet overeenkomen met de parameter \code{dim} in
\code{\link{addDimView}}}

\item{skipTopRow}{boolean, Als TRUE: eerste regel met Naam, links en presentatie wordt overgeslagen.}

\item{maxHeight}{an integer, maximum hoogte in pixels voor dit dimView object}

\item{overflowX}{string, css overflow propery in X richting, bepaalt of er wel of niet een horizontale scrollbar getoond wordt bij overflow.}
}
\description{
Creeer de html voor het dimView object in de shiny app's UI.
}
